+++ src/App.tsx
@@ -1,17 +1,12 @@
-import { useEffect, Suspense } from "react";
+import { Suspense } from "react";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
+import { useTranslation } from "react-i18next";
+import clsx from "clsx";
 import { Layout, Loading } from "@components";
 import { Home, Note, Archive, Trash, NotFound, Login } from "@pages";
-import clsx from "clsx";
-import { useTranslation } from "react-i18next";
-import { initialThemeMode } from "@utils";
 
 export default function App(): JSX.Element {
   const { i18n } = useTranslation();
-
-  useEffect(() => {
-    initialThemeMode();
-  }, []);
 
   return (
     <div className={clsx("dark:text-white", i18n.language)}>