+++ src/utils/index.ts
@@ -4,6 +4,7 @@ export {
   LOCAL_STORAGE_TOKEN,
   LOCAL_STORAGE_USER,
 } from "./constants";
+
 export { getQueryString } from "./routerUtils";
-export { isDarkMode, initialThemeMode, getDocumentTitle } from "./styleUtils";
+export { getDocumentTitle } from "./styleUtils";
 export { isLogin, getUserInfoFromStorage, getToken } from "./userUtils";