+++ src/utils/styleUtils.ts
@@ -1,21 +1,4 @@
-import { APPLICATION_NAME } from "./constants";
-
-function isDarkMode(): boolean {
-  return document.documentElement.classList.contains("dark");
-}
-
-function initialThemeMode() {
-  const darkModeFromStorage: string | null =
-    window.localStorage.getItem("theme");
-  switch (darkModeFromStorage) {
-    case "dark":
-      document.documentElement.classList.add("dark");
-    case "light":
-    case null:
-    default:
-      return;
-  }
-}
+import { APPLICATION_NAME } from "@utils";
 
 function getDocumentTitle(prefix: string): string {
   if (!prefix) {
@@ -24,4 +7,4 @@ function getDocumentTitle(prefix: string): string {
   return prefix + " | " + APPLICATION_NAME;
 }
 
-export { isDarkMode, initialThemeMode, getDocumentTitle };
+export { getDocumentTitle };