+++ src/utils/userUtils.ts
@@ -2,12 +2,12 @@ import { LOCAL_STORAGE_TOKEN, LOCAL_STORAGE_USER } from "@utils";
 import { type UserInfo } from "@store";
 
 function isLogin(): boolean {
-  const hasToken = window.localStorage.getItem(LOCAL_STORAGE_TOKEN);
+  const hasToken = localStorage[LOCAL_STORAGE_TOKEN];
   return hasToken !== null && hasToken !== "";
 }
 
 function getUserInfoFromStorage(): UserInfo | null {
-  let userData = window.localStorage.getItem(LOCAL_STORAGE_USER);
+  let userData = localStorage[LOCAL_STORAGE_USER];
   if (!userData) {
     return null;
   }
@@ -16,7 +16,7 @@ function getUserInfoFromStorage(): UserInfo | null {
 }
 
 function getToken(): string | null {
-  return window.localStorage.getItem(LOCAL_STORAGE_TOKEN);
+  return localStorage[LOCAL_STORAGE_TOKEN];
 }
 
 export { isLogin, getUserInfoFromStorage, getToken };